containers/guacamole/controller/src/requirements.txt
click==8.1.7
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
requests==2.31.0
python-ldap==3.4.4
